requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
Django==4.2.10
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.49.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-vision==3.7.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
import-export==0.3.1
importlib-resources==6.1.1
ipython==8.18.1
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
numpy==1.26.4
openai==1.12.0
opencv-contrib-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
utils==1.0.2
wcwidth==0.2.13
zipp==3.17.0
